feat: make virtual chunk allocator bigint-native for All BTC keyspace (#79)
- Add alloc_cursor_hex / virtual_chunk_count_hex TEXT columns to puzzles
- Add vchunk_start_hex / vchunk_end_hex TEXT columns to chunks (64-char
  zero-padded hex enables correct SQL lexicographic comparison)
- Backfill migration: printf('%064x') converts existing INTEGER rows
- Remove INT64_MAX guard from chooseDefaultVirtualChunkSize() — chunk
  size no longer inflated just to make count fit in int64_t
- Remove INT64_MAX throw from seedVirtualChunks(); store large counts
  as hex; keep INTEGER column NULL when value exceeds int64_t range
- Change PuzzleRow::allocCursor/virtualChunkCount from int64_t to
  cpp_int; readPuzzle() reads hex columns with fallback to INTEGER
- Port all virtual-domain variables (totalChunks, cursor, orderIndex,
  candidateIndex, runStart, runEnd, neededChunks, bootstrap runs) to
  cpp_int throughout the allocator
- rangeIsFree() now queries vchunk_start_hex/vchunk_end_hex TEXT columns
- setAllocCursor() writes alloc_cursor_hex; assignVirtualChunkRun()
  writes both hex TEXT and INTEGER (NULL when too large) columns
- service_serialization: output alloc_cursor/virtual_chunk_count as
  decimal strings to avoid JSON number precision loss
- 5 new test cases: large-domain seeding, hex persistence, NULL INTEGER,
  correct package sizing at 100 MH/s, non-overlapping assignment

+++ include/puzzpool/types.hpp
@@ -30,9 +30,9 @@ struct PuzzleRow {
     std::string testEndHex;
     std::string allocStrategy;
     std::string allocSeed;
-    int64_t allocCursor = 0;
+    cpp_int allocCursor = 0;
     std::string virtualChunkSizeKeys;
-    int64_t virtualChunkCount = 0;
+    cpp_int virtualChunkCount = 0;
     int bootstrapStage = 0;
 };
